Sync template (#32)
* Update INSTALL.md (#3074)

Again, I saw this version problem with current version of our docker
image. Supposedly, there is some json package problem which I dealt
with.

But here is the overall solution to this problem for whoever is having
it.

---------

Co-authored-by: George <31376482+george-gca@users.noreply.github.com>

* Re-add ability to exclude posts from search (#3089)

Earlier there was a tag in the `_config.yml` file to exclude the posts
from the search results. I could no longer find that option and hence
modified the `search.liquid.js` file.
It basically checks if the `posts_in_search` tag is true or not.

* Updated dependencies (#3091)

Signed-off-by: George Araújo <george.gcac@gmail.com>

* Added support for plotly js (#3097)

Only now I realized that the [previous
support](https://alshedivat.github.io/al-folio/blog/2021/distill/#interactive-plots)
for [plotly](https://plotly.com/) was enabled by exporting an HTML from
python code. This adds support for
[plotly.js](https://plotly.com/javascript/).

---------

Signed-off-by: George Araújo <george.gcac@gmail.com>

* Fix "comma and and" before last author in publication list (#3107)

* Fixed liquid variables in some js files (#3117)

Fixes #3095

---------

Signed-off-by: George Araújo <george.gcac@gmail.com>

* Changed broken-links action to only run on main repo (#3118)

Signed-off-by: George Araújo <george.gcac@gmail.com>

---------

Signed-off-by: George Araújo <george.gcac@gmail.com>
Co-authored-by: Amir Pourmand <pourmand1376@gmail.com>
Co-authored-by: George <31376482+george-gca@users.noreply.github.com>
Co-authored-by: Pranjal Aggarwal <54495980+pranjalagg@users.noreply.github.com>
Co-authored-by: Thomas Kahle <thomas.kahle@jpberlin.de>